backend/requirements.txt
fastapi>=0.116.0,<0.118.0
uvicorn[standard]>=0.30.0
sqlalchemy==2.0.35
groq>=0.11.0
pandas==2.2.3
scikit-learn==1.5.2
python-dotenv==1.0.1
pydantic>=2.11.0
aiofiles>=24.1.0
python-multipart>=0.0.9
